backend/requirements.txt

fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Google Drive API
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0

# Database
sqlite3-binary ; sys_platform != "win32"

